server/requirements.txt
atomicwrites==1.3.0
attrs==18.2.0
Click==7.0
coverage==4.5.2
Flask==1.0.2
Flask-Cors==3.0.7
Flask-Script==2.0.6
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
mongoengine==0.16.3
more-itertools==6.0.0
pkg-resources==0.0.0
pluggy==0.8.1
py==1.7.0
pymongo==3.7.2
pytest==4.2.1
python-dotenv==0.10.1
pytils==0.3
six==1.12.0
Unidecode==1.0.23
Werkzeug==0.14.1
